add gateway
