+++ PoC2/gateway/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+
+def create_app(config_name):
+    app = Flask(__name__)
+   
+    return app
