+++ PoC2/requirements.txt
@@ -24,7 +24,7 @@ marshmallow==3.12.1
 marshmallow-sqlalchemy==0.25.0
 mccabe==0.6.1
 prompt-toolkit==3.0.18
-psycopg2==2.8.6
+psycopg2-binary==2.9.1
 PyJWT==2.1.0
 pylint==2.8.3
 pytz==2021.1
